feat (rendering) : adds basic lighting
To bring a bit of depth to the world. Added normals to the chunk mesh by passing a 'normal index' to the gpu as a single byte. The vertex shader then select the proper normal according the the index(There are only 6 possible normals since faces are aligned with the x,y and z axis).